BFGS lpdf

A wrapper for codeBFGS_std that is useful for easily calling parameter optimization for this package with as few lines as possible. Note that om and logpdf will be set to optimal parameters, the return is simply for information.

Usage

BFGS_lpdf(
  om,
  logpdf,
  par = list(),
  newt = F,
  cgsteps = 100,
  cgtol = 0.001,
  ...
)

Arguments

om

an outermod object

logpdf

a lpdf object

par

an initial point, initialized from objects if needed

newt

boolean for if Newtons method should be used

cgsteps

max number of cg iterations, if newt=FALSE

cgtol

cg tolerance, if newt=FALSE

...

additional parameters passed to BFGS_std

Value

A list of information from optimization